correct histogram code sample
