+++ elements/charts/bar_chart.py
@@ -10,23 +10,6 @@ def run():
     with st.echo():
 
         @st.cache
-        def get_histo():
-
-            df = pd.DataFrame(
-                    np.random.randn(200, 1),
-                    columns=['a']
-                )
-
-            return np.histogram(
-                df.a, bins=25)[0]
-
-        hist_values = get_histo()
-
-        st.bar_chart(hist_values)
-
-    with st.echo():
-
-        @st.cache
         def get_bar_chart_data():
 
             return pd.DataFrame(
