+++ elements/charts/order.json
@@ -1,5 +1,6 @@
 [
   "line_chart",
   "area_chart",
+  "histogram",
   "bar_chart"
 ]
